cwt_agent/requirements.txt
hermes-agent>=0.1.0
openai>=1.30.0          # OpenRouter is OpenAI-compatible
praw>=7.7.1
apify-client>=1.7.0
python-dotenv>=1.0.0
requests>=2.31.0
beautifulsoup4>=4.12.3
lxml>=5.1.0
rich>=13.7.0            # pretty terminal output
tenacity>=8.2.3         # retry logic
